report: parse integers in range
This patch adds the parsing of the integers specified in the range part
of each field of the group-by argument.

Also change the usage of strtok in the group-by field parsing to
strtok_r.

+++ librobinhood/include/robinhood/backend.h
@@ -159,7 +159,7 @@ struct rbh_filter_options {
 
 struct rbh_range_field {
     struct rbh_filter_field field;
-    int *boundaries;
+    int64_t *boundaries;
     size_t boundaries_count;
 };
